ref(build): Programmatically generate base CDN rollup config (#4650)
As part of the new build process, this continues the work centralizing the rollup config for our CDN bundles. More specifically, this creates a function to generate the base configs for `@sentry/tracing`, `@sentry/vue`, and `@sentry/wasm`, and changes each of their respective rollup configs to use it. As expected with a pure refactor, there are no bundle changes.

Not included in this PR are `@sentry/browser` and `@sentry/integrations`. Because their respective rollup configs are more complicated, they'll be handled in future PRs. As a result, code which will, once those are handled, replace much of the code in the repo-level `rollup.config.js` has instead been added, so as to not break their current build processes.

Note that the added dev dependency doesn't actually change `yarn.lock`, because `deepmerge` is already included, as a transitive dependency.