ref(utils): Improve logging of objects in `logger` methods (#4663)
When a variable is used in a template string, its `.toString()` method is called, and most non-native objects, when stringified with `.toString()`, come out as `"[object Object]"`. In our `logger`, we use a template string to concatenate all of the arguments we're given. As a result, users can get error messages like:

```
Sentry Logger [Error]: Error while sending event: [object Object]
```

in cases where the error is not an actual `Error`, which is obviously not very helpful.

The underlying `console` methods upon which our `logger` methods rely can generally do better, though, and will attempt to print the actual contents of any object they're given. This changes the `logger` methods so that they pass their arguments on to the `console` methods intact, rather than as a pre-concatenated string, in order to be able to take advantage of the improved logging of objects. It also changes all instances of "error while" logs like the one above to pass their error objects separately rather than as part of a template string, in order to take advantage of this change.

Note: There are almost certainly other places in the repo where we log objects which could benefit from this change. In the interests of time, however, this PR is restricted to the ones which brought this issue to light, and others can be solved in future PRs as necessary.