fix(build): Build bundles in parallel (#4678)
This parallelizes our top-level `build:bundle` yarn script, in order to make it run faster. Because so much of the total time is spent on one package (`@sentry/integrations`, which currently produces 24 bundles to every other package's max 4 bundles), the time savings isn't earth-shattering, but it does run 20-30% faster in my informal tests.