feat(tracing): Add ES6 tracing bundle (#4674)
This adds an ES6 version of our tracing CDN bundle, both for the sake of users and as a step towards becoming ES6-first in v7. It also adds the new bundle to the size check and the playwright integration tests.

A note about the new bundle is added to the docs in https://github.com/getsentry/sentry-docs/pull/4789.