fix(nextjs): Add env var to suppress API non-response meta-warning (#4706)
In the nextjs SDK, when we wrap users' API routes, we also wrap the response's `end` method, in order to keep the lambda running the route handler alive long enough to send events to Sentry. As a consequence, however, Next thinks a response hasn't been sent at all, because `end` hasn't been called within the timeframe that it expects, so it throws a warning in dev. 

A previous attempt[1] to fix this problem backfired[2], and had to be reverted[3], so as a compromise option for the moment, we log a warning about the Next warning, so at least people know it's nothing to worry about. Some people are finding this behavior spammy[4], though, so this PR adds an env variable check which allows a user to suppress our meta-warning, and also improves the warning itself so that people know the option is there.

[1] https://github.com/getsentry/sentry-javascript/pull/4139
[2] https://github.com/getsentry/sentry-javascript/issues/4151
[3] https://github.com/getsentry/sentry-javascript/pull/4516
[4] https://github.com/getsentry/sentry-javascript/issues/3852#issuecomment-1062933254