ref(tracing): Remove `BrowserTracing` logging flag default value (#4708)
In the `BrowserTracing` integration, we use a flag as a way to delay logging a warning message[1] from the constructor until we can guarantee the logger has been enabled. Until now, that flag defaulted to `false`, which has prevented it from being treeshaken, even in non-debug bundles in which it's not otherwise touched.

This PR makes the flag optional, and removes the default value. Since `undefined` is falsy there are no runtime behavior changes, but the fact that it's never used, not even to be set to a default value, allows it to be fully removed by treeshaking. 

[1] https://github.com/getsentry/sentry-javascript/blob/af7081c9e9be240578cf608048734572ce709ac0/packages/tracing/src/browser/browsertracing.ts#L136-L144