fix(dev): Use correct schema for GHA workflow files in VSCode (#4727)
VSCode has a feature[1] wherein it auto-detects what schema to use to validate certain JSON and yml files. The problem is, sometimes it's wrong[2], and when it is, it will throw errors* because it's misconstrued what kind of file you're editing. The incorrect association can be overridden, though, with a correct association in settings.

This adds such an association, to prevent VSCode from complaining about our GHA `build` workflow.

*It correctly detects that it's a GHA config, but also checks it against the schema for a different build tool[3] called `hammerkit`, and throws errors because the two schemas are incompatible

[1] https://code.visualstudio.com/docs/languages/json#_json-schemas-and-settings
[2] https://github.com/microsoft/vscode/issues/74943#issuecomment-876013151
[3] https://no0dles.gitbook.io/hammerkit/