chore(various): Autoformat JS files (#4726)
This applies the autoformatter to all of the JS (as in, written originally in JS, not transpiled into JS) files in the repo, to avoid having future changes to those files polluted by irrelevant autoformating triggered by saving the file. A few non-JS files, primarily in the ember package, are also included.