fix(dev): Add missing `--ignore` in main `yarn test` command (#4756)
As part of https://github.com/getsentry/sentry-javascript/pull/4694, which adds integration tests for node, a change was made to the repo-level `yarn test` command, in order to skip running the new tests. The new version of the command was missing a `--ignore` flag, though, causing the command to error out. This adds the missing flag.