feat(core): Add new transports to base backend (#4752)
Adds new transports to base backend in core. For now, they are gated behind `options._experiments.newTransport = true`. The reason this is gated is because client reports does not work with the new transports.

The next step is to add new transports for fetch, xhr (browser) as well as http, https (node). We can do this in any order!