feat(vue): Drop vue-router peerDep (#4764)
We previously added vue-router as a peer dependency because we rely on
it's existence for Vue Performance Monitoring's router instrumentation.
We never import directly from this though, we always rely on the user
passing it in afterwards. As such, we don't really need to define
a peerDep for the vue-router package.

We use a similar pattern for the react-router stuff, and it has worked
out super well.