fix(integrations): Fix type/async errors in `Offline` tests (#4755)
This fixes a number of longstanding types errors in the tests for the `Offline` integration. It also simplifies some of the async code (to avoid mixing async functions and `done()`) and changes a mock so it is no longer assigning to a read-only property.

Getting these fixed will help unmuddy the waters with respect to the `Offline` test errors which are cropping up in our TS upgrade.