ref(utils): Use type predicates for `isError` and `isString` (#4757)
This restores part of the work done in https://github.com/getsentry/sentry-javascript/pull/4124, which then got reverted in https://github.com/getsentry/sentry-javascript/pull/4149. The problem there was the predicates for `DOMError` and `DOMException`, both of which only exist in a browser context. This restores the predicates for `isError` and `isString`, which should be safe, because both `Error` and `string` exist in both browser and node contexts.