fix(core): Use correct version of event when tagging normalization (#4780)
In https://github.com/getsentry/sentry-javascript/pull/4425, various hacks were added to the SDK to try to diagnose the root cause of https://github.com/getsentry/sentry-javascript/issues/2809. Unfortunately, one of those hacks (tagging when events are normalized) was applied to the pre-normalized version of the event rather than the normalized version that continues on through the event processing pipeline. This led to false positives, where events got tagged as if they'd skipped the normalization process even when they hadn't.

This fixes that by adding the flag to the normalized version of the event instead.