fix(core): Stop mangling `_experiments` (#4807)
We hide experimental options behind an `_experiments` flag. This is fine for folks using our npm packages, but for folks using CDN bundles it's a problem, because until now, we've been mangling the property name `_experiments`. (Normally, properties starting with a single underscore are private class methods or fields, the names of both of which it's fine to mangle. We therefore use that as our criterion when telling terser what it may and may not touch, and `_experiments` has been getting swept up by that regex.)

This carves out an exception for `_experiments`, so that it is no longer mangled.