fix(types): Remove unnecessary type assertions (#4821)
Now that most of our `isXXX` functions use type predicates, there are a number of places in the code where we no longer have to cast values to a certain type. This covers a few of those instances (pulled from another PR to reduce noise), along with a few others the linter found.