ref(build): Add `build/types` to tarballs and adjust `types` entry points (#4824)
* All tarballs now include the `build/types` directory. All `.npmignore`s were adjusted to include this directory.
* The `types` entry point in all `package.json`s are now adjusted to the new types directory

Proper bugfix for 6.19.0 regression:
With PR #4724, we introduced separate building of type declaration files which would be written to `<sdk>/build/types`. Furthermore, we adjusted the `types` entry point in the `package.json`s to this new directory. To avoid a breaking change, we kept type declarations in `dist` and `esm`. However, a bug in this PR caused a small regression in most of our SDKs: The new types directory was not included in most of our SDK tarballs. Subsequently, we released a hot fix via #4745 which simply reset the entry points back to their previous path (`dist/index.d.ts`) in all SDKs. 

Additional adjustments w.r.t. creating tarballs to improve consistency:
* Browser: Adjusted `postbuild.sh` to copy `.npmignore` to `./build`
* Gatsby: Use `.npmignore` instead of `files` property in `package.json`
* Utils: Changed `.npmignore` to be more similar to our other packages
* WASM: Changed `.npmignore` to be more similar to our other packages
* Next.js: Added `.npmignore`