ref(build): convert `postbuild.sh` to node script
add `postbuild.ts` node script which can be run with ts-node to replace the
bash script `postbuild.sh`. This is done to conform with our decision to
change all bash scripts to node s.t. they can be run on any platform (incl.
Windows). Furthermore, it gives us the flexibility to make more adjustments
that are harder to code/read in bash.

* clone behaviour of `postbuild.sh`
* add removal of `volta` and `scripts` properties in copied package.json
* adjust calls to postbuild.sh to node script
* adjust reference in comment to node script