ref(dev): Run linter on scripts (#4834)
Because SDK developers running Windows [can't run our bash scripts[1], we recently decided to move to using TypeScript instead. To make this easier, this extends the domain of our `eslint` and `prettier` checks to include all `scripts` folders. It also fixes a few linting errors which doing so brought to light.

[1] https://github.com/getsentry/sentry-javascript/issues/4720