fix(build): Raise time limit for build to 20 minutes (#4847)
All of the changes happening to the build process (and all of the new bundles we're producing) mean that sometimes the build step is failing because it times out. While in the long run we should fix the build to be faster, in the short run, this ups the time limit, so at least CI won't error out unnecessarily.