ref(build): Convert `postbuild.sh` to node script (#4828)
Replace the bash script `postbuild.sh` with the node script `postbuild.ts` which can be run with `ts-node`. This is done to conform with our decision to change all bash scripts to node s.t. they can be run on any platform (incl. Windows). Furthermore, it gives us the flexibility to add additional functionality in the future that is harder to code/read in bash.

Co-authored-by: Katie Byers <lobsterkatie@gmail.com>