feat(utils): Introduce getGlobalSingleton helper
This helper abstracts away the logic of getting an variable from the
global `__SENTRY__` object. It does this by introducing a new
`getGlobalSingleton` utility function.

For now, this is used by the event processors and hub logic. In an
upcoming patch, we'll update the logger logic to also use this utility.