ref(core): Log `normalizeDepth` type when normalization is skipped (#4870)
This is a follow up to https://github.com/getsentry/sentry-javascript/pull/4425 and https://github.com/getsentry/sentry-javascript/pull/4574, all in aid of debugging https://github.com/getsentry/sentry-javascript/issues/2809, wherein serializing events leads to a circular reference error. One of the possible culprits is that the event is somehow skipping our usual normalization process, and one of the reasons that might happen is if `normalizeDepth` is set to something other than a number. We're already logging its value in problematic cases, and this change makes it so that we'll now log its type as well.