poc(perf): Add browser session id to transactions
An example of how we might add browser sessions (via session-storage) to experiment with performance across user journies. We can release this as a private option, or try out this branch via beta on the Sentry app. Use of sessionStorage vs. localstorage is debatable and likely something we'd need to experiment with.