ref: Inject Transports into Client (#4921)
Changes the way we initialize Transports. Previously, the Transport was initialized within the Client constructor (by calling setupTransport()). Change this so that the Transport is initialized beforehand and then injected into the client. This is currently (i.e. with this PR) happening via two additional arguments in the Client class constructors. The reason for two arguments is that we're still using both, the old `Transport` classes and the `NewTransport` interface. In the future, `NewTransport` will replace `Transport` and (once the old Transport is removed) the transport is then passed into the Client constructor as a property of the options object.

* add the injection logic
* initialize the transports in Browser and Node and pass them to the `Browser/NodeClient` initialization
* add client-external transport setup functions (extracted from setupTransport())
* add basic tests for these transport setup functions
* delete the client-internal setupTransport methods
* fixe a bunch of tests that are initializing clients