doc(migration): Add `Backend` removal and `Transport` injection to migration docs (#4934)
Add two small bullet points to the "General API changes" section, briefly explaining the removal of `Backend` and the injection of `Transport`s. The transport injection part needs to be revisited once we switch over to just using `NewTransport` as well as when we're passing the transports as `options` properties to the client constructors.