chore(gatsby): Create type declarations for gatsby plugin files (#4928)
When creating a Gatsby plugin (which is what `@sentry/gatsby` actually is), there are certain files[1] which need to live at the package root level. Because they are written in JS (not TS) and don't live in `src/`, we have thus far more or less ignored them as far as type-ful processing goes. With the recent updates to TS, jest, and ts-jest, typechecking has now gotten stricter (especially in tests) and so now declaration files are needed for those gatsby-required, root-level files if we want to test them (which we do).

This adds machinery to create, manage, and publish these declaration files.

[1] https://www.gatsbyjs.com/docs/files-gatsby-looks-for-in-a-plugin/