fix(build): Cache gatsby plugin types files in CI (#4944)
This is a follow up to https://github.com/getsentry/sentry-javascript/pull/4928, which added the creation of types files for the code in the gatsby SDK which lives outside of `src/`. One change which was missed was to have GHA include these new files in the build cache, so that they are available when tests are run. This fixes that oversight.