fix(dev): Check out correct commit in manual GHA `Build & Test` workflow (#4954)
Our `Build & Test` GHA workflow can be triggered manually, and when you trigger it, you can either pick the branch to run it on or give it a specific commit to test. The former works fine, but the latter does not - though it picks up the given commit and correctly puts it in the env, it doesn't actually use that value when checking out the commit, instead always relying on the default value. This fixes that by always specifying what value to use.