feat(angular): Log warning if SDK is used with an older Angular version than officially supported (#4964)
introduces a check in the Angular SDK's init function to check if the Angular version it is used with is older than our minimum supported version (Angular 10). In case it is, a warning will be logged.

adjusts the migration and readme docs to reflect better our officially supported Angular versions and what to do if users are using an older Angular version if they experience problems (i.e. stay on v6).


Co-authored-by: Luca Forstner <luca.forstner@sentry.io>