ref: Switch to new transports (#4943)
This PR switches over from the old transports to the new ones. It deletes the functionality from the client of passing in transports explicitly and instead expects the transport to be passed in through options.

Instead of passing in an instance of a transport to the client, we pass in a constructor function. This allows the client to control exactly what options are passed into the transport.

```ts
this._transport = options.transport({ ...options.transportOptions, url });
```