fix(tests): Silence dummy output in tests (#4979)
We have a number of tests which log messages to the console, for reasons other than communicating with the person running the tests. When trying to debug a large number of tests, across the whole repo, it's distracting to see this output (some of which actually even looks like failing tests). This silences that output, so it doesn't seem like tests are erroring when they're not.