feat(hub): Remove _invokeClient (#4972)
This PR removes the dynamic dispatch of `_invokeClient` in favour of explicitly using client methods. We introduce a `_withClient` helper that does this.

`_withClient` grabs the client and scope from the stack of scopes on the hub, and executes a callback with those values. This allows the consumer of the function to directly access the client instance and call methods on it.