feat: Add `nodeId` when capturing LCP (getsentry/sentry-replay#60)
The PerformanceObserver allows us to observe LCP events. The LCP performance entry has a reference to the element in the DOM that caused the LCP event. We can use this element reference to query rrweb to get the serialized rrweb metadata -- relevant to us is the nodeId. This will allow our replayer to do the inverse query and get a reference to the element in the replayer via nodeId so that we can "highlight" it.